chore(store): move types and re-exports out of widget store
